requirements.txt
fastapi
uvicorn
PyPDF2
langchain
langchain_google_genai
faiss-cpu
python-dotenv
google-generativeai
langchain_community
python-multipart
pytest 
httpx